backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Settings
python-dotenv==1.0.0